refactor gateway code to accept otto or goja as jsvm
