+++ mw_virtual_endpoint_test.go
@@ -11,6 +11,7 @@ import (
 
 const virtTestJS = `
 function testVirtData(request, session, config) {
+
 	var resp = {
 		Body: "foobar",
 		Headers: {
@@ -19,6 +20,7 @@ function testVirtData(request, session, config) {
 		},
 		Code: 202
 	}
+	console.log(TykJsResponse(resp, session.meta_data))
 	return TykJsResponse(resp, session.meta_data)
 }
 `
